+++ NETCORE/test/Microsoft.ApplicationInsights.AspNetCore.Tests/Extensions/ApplicationInsightsServiceOptionsTests.cs
@@ -1,17 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Text;
 
 using Microsoft.ApplicationInsights.AspNetCore.Extensions;
+using Microsoft.ApplicationInsights.Channel;
 using Microsoft.ApplicationInsights.DependencyCollector;
 using Microsoft.ApplicationInsights.Extensibility;
 #if NETCOREAPP
 using Microsoft.ApplicationInsights.Extensibility.EventCounterCollector;
-#endif 
+#endif
+using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing;
 using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector;
 using Microsoft.ApplicationInsights.Extensibility.PerfCounterCollector.QuickPulse;
 using Microsoft.ApplicationInsights.WindowsServer;
@@ -19,6 +18,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Hosting.Internal;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Test;
 
 using Xunit;
 
@@ -39,7 +39,7 @@ namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Extensions
     /// </remarks>
     public class ApplicationInsightsServiceOptionsTests
     {
-        private static IServiceProvider TestShim(string configType, bool isEnabled, Action<ApplicationInsightsServiceOptions, bool> testConfig)
+        private static IServiceProvider TestShim(string configType, bool isEnabled, Action<ApplicationInsightsServiceOptions, bool> testConfig, Action<IServiceCollection> servicesConfig = null)
         {
             // ARRANGE
             Action<ApplicationInsightsServiceOptions> serviceOptions = null;
@@ -55,8 +55,9 @@ namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Extensions
 
             // ACT
             var services = CreateServicesAndAddApplicationinsightsWorker(
-                jsonPath: filePath, 
-                serviceOptions: serviceOptions, 
+                jsonPath: filePath,
+                serviceOptions: serviceOptions,
+                servicesConfig: servicesConfig,
                 useDefaultConfig: configType == "DefaultConfiguration" ? true : false);
 
             IServiceProvider serviceProvider = services.BuildServiceProvider();
@@ -71,7 +72,7 @@ namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Extensions
             return serviceProvider;
         }
 
-        private static ServiceCollection CreateServicesAndAddApplicationinsightsWorker(string jsonPath, Action<ApplicationInsightsServiceOptions> serviceOptions = null, bool useDefaultConfig = true)
+        private static ServiceCollection CreateServicesAndAddApplicationinsightsWorker(string jsonPath, Action<ApplicationInsightsServiceOptions> serviceOptions = null, Action<IServiceCollection> servicesConfig = null, bool useDefaultConfig = true)
         {
             IConfigurationRoot config;
             var services = new ServiceCollection()
@@ -114,12 +115,75 @@ namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Extensions
             }
 #endif
 
+            servicesConfig?.Invoke(services);
+
             if (serviceOptions != null)
             {
                 services.Configure(serviceOptions);
             }
 
             return (ServiceCollection)services;
+        }
+
+        /// <summary>
+        /// User could enable or disable TelemetryConfiguration.Active by setting EnableAppServicesHeartbeatTelemetryModule.
+        /// </summary>
+        /// /// <summary>
+        /// This SDK previously had a hidden dependency on TelemetryConfiguration.Active.
+        /// We've removed that, but users may have taken a dependency on the former behavior.
+        /// This test verifies that users can enable "backwards compat".
+        /// Enabling this will copy the AspNetCore config to the TC.Active static instance.
+        /// </summary>
+        [Theory]
+#if !NET46
+        [InlineData("DefaultConfiguration", true)]
+        [InlineData("DefaultConfiguration", false)]
+        [InlineData("SuppliedConfiguration", true)]
+        [InlineData("SuppliedConfiguration", false)]
+#endif
+        [InlineData("Code", true)]
+        [InlineData("Code", false)]
+        public static void UserCanEnableAndDisableTelemetryConfigurationActive(string configType, bool isEnable)
+        {
+#pragma warning disable CS0618 // Type or member is obsolete
+            // Dispose .Active to force a new .Active to be created during this test.
+            TelemetryConfiguration.Active.Dispose();
+
+            // IMPORTANT: This is the same ikey specified in the config files that will be used for this test.
+            string testString = "22222222-2222-3333-4444-555555555555";
+            var testTelemetryInitializer = new FakeTelemetryInitializer();
+
+            IServiceProvider serviceProvider = TestShim(
+                configType: configType,
+                isEnabled: isEnable,
+                testConfig: (o, b) =>
+                {
+                    o.EnableActiveTelemetryConfigurationSetup = b;
+                    o.InstrumentationKey = testString;
+                },
+                servicesConfig: (services) => services.AddSingleton<ITelemetryInitializer>(testTelemetryInitializer)
+                );
+
+            // TelemetryConfiguration from DI should have custom set InstrumentationKey and TelemetryInitializer
+            var telemetryConfiguration = serviceProvider.GetTelemetryConfiguration();
+            Assert.Equal(testString, telemetryConfiguration.InstrumentationKey);
+            Assert.Same(testTelemetryInitializer, telemetryConfiguration.TelemetryInitializers.OfType<FakeTelemetryInitializer>().Single());
+
+            // TelemetryConfiguration.Active will only have custom set InstrumentationKey if .Active was enabled.
+            Assert.Equal(testString.Equals(TelemetryConfiguration.Active.InstrumentationKey), isEnable);
+            
+            // TelemetryConfiguration.Active will only have custom TelemetryInitializer if .Active was enabled
+            var activeTelemetryInitializer = TelemetryConfiguration.Active.TelemetryInitializers.OfType<FakeTelemetryInitializer>().SingleOrDefault();
+            if (isEnable)
+            {
+                Assert.Same(testTelemetryInitializer, activeTelemetryInitializer);
+            }
+            else
+            {
+                Assert.Null(activeTelemetryInitializer);
+            }
+
+#pragma warning restore CS0618 // Type or member is obsolete
         }
 
         /// <summary>
@@ -136,7 +200,7 @@ namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Extensions
         [InlineData("Code", false)]
         public static void UserCanEnableAndDisablePerfCollectorModule(string configType, bool isEnable)
         {
-            IServiceProvider serviceProvider = TestShim(configType: configType, isEnabled: isEnable, testConfig: (o, b) => o.EnablePerformanceCounterCollectionModule = b );
+            IServiceProvider serviceProvider = TestShim(configType: configType, isEnabled: isEnable, testConfig: (o, b) => o.EnablePerformanceCounterCollectionModule = b);
 
             var modules = serviceProvider.GetServices<ITelemetryModule>();
             var module = modules.OfType<PerformanceCollectorModule>().Single();
@@ -269,6 +333,51 @@ namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Extensions
             var modules = serviceProvider.GetServices<ITelemetryModule>();
             var module = modules.OfType<AppServicesHeartbeatTelemetryModule>().Single();
             Assert.Equal(isEnable, module.IsInitialized);
+        }
+
+        /// <summary>
+        /// User could enable or disable <see cref="DiagnosticsTelemetryModule"/> by setting <see cref="ApplicationInsightsServiceOptions.EnableDiagnosticsTelemetryModule"/>.
+        /// </summary>
+        [Theory]
+#if !NET46
+        [InlineData("DefaultConfiguration", true)]
+        [InlineData("DefaultConfiguration", false)]
+        [InlineData("SuppliedConfiguration", true)]
+        [InlineData("SuppliedConfiguration", false)]
+#endif
+        [InlineData("Code", true)]
+        [InlineData("Code", false)]
+        public static void UserCanEnableAndDisableDiagnosticsTelemetryModule(string configType, bool isEnable)
+        {
+            IServiceProvider serviceProvider = TestShim(configType: configType, isEnabled: isEnable, testConfig: (o, b) => o.EnableDiagnosticsTelemetryModule = b);
+
+            var modules = serviceProvider.GetServices<ITelemetryModule>();
+            var module = modules.OfType<DiagnosticsTelemetryModule>().Single();
+            Assert.Equal(isEnable, module.IsInitialized);
+        }
+
+        /// <summary>
+        /// User could enable or disable the Heartbeat feature by setting <see cref="ApplicationInsightsServiceOptions.EnableHeartbeat"/>.
+        /// </summary>
+        /// <remarks>
+        /// Config file tests are not valid in this test because they set ALL settings to either TRUE/FALSE.
+        /// This test is specifically evaluating what happens when the DiagnosticsTelemetryModule is enabled, but the Heartbeat feature is disabled.
+        /// </remarks>
+        [Theory]
+        [InlineData("Code", true)]
+        [InlineData("Code", false)]
+        public static void UserCanEnableAndDisableHeartbeatFeature(string configType, bool isEnable)
+        {
+            IServiceProvider serviceProvider = TestShim(configType: configType, isEnabled: isEnable,
+                testConfig: (o, b) => {
+                    o.EnableDiagnosticsTelemetryModule = true;
+                    o.EnableHeartbeat = b;
+                });
+
+            var modules = serviceProvider.GetServices<ITelemetryModule>();
+            var module = modules.OfType<DiagnosticsTelemetryModule>().Single();
+            Assert.True(module.IsInitialized, "module was not initialized");
+            Assert.Equal(isEnable, module.IsHeartbeatEnabled);
         }
     }
 }